+++ src/mongo.c
@@ -154,23 +154,17 @@ static int mongo_check_is_master( mongo *conn ) {
 }
 
 void mongo_init( mongo *conn ) {
-    conn->replset = NULL;
-    conn->err = 0;
-    conn->errstr = NULL;
-    conn->lasterrcode = 0;
-    conn->lasterrstr = NULL;
-
-    conn->conn_timeout_ms = 0;
-    conn->op_timeout_ms = 0;
+    memset( conn, 0, sizeof( mongo ) );
 }
 
 int mongo_connect( mongo *conn , const char *host, int port ) {
+    mongo_init( conn );
+
     conn->primary = bson_malloc( sizeof( mongo_host_port ) );
     strncpy( conn->primary->host, host, strlen( host ) + 1 );
     conn->primary->port = port;
     conn->primary->next = NULL;
 
-    mongo_init( conn );
     if( mongo_socket_connect( conn, host, port ) != MONGO_OK )
         return MONGO_ERROR;
 
@@ -374,8 +368,11 @@ int mongo_replset_connect( mongo *conn ) {
                     return MONGO_ERROR;
 
                 /* Primary found, so return. */
-                else if( conn->replset->primary_connected )
+                else if( conn->replset->primary_connected ) {
+                    strncpy( conn->primary->host, node->host, strlen( node->host ) + 1 );
+                    conn->primary->port = node->port;
                     return MONGO_OK;
+                }
 
                 /* No primary, so close the connection. */
                 else {
@@ -494,12 +491,14 @@ static int mongo_bson_valid( mongo *conn, bson *bson, int write ) {
 /* Determine whether this BSON object is valid for the given operation.  */
 static int mongo_cursor_bson_valid( mongo_cursor *cursor, bson *bson ) {
     if( ! bson->finished ) {
-        cursor->err = MONGO_BSON_NOT_FINISHED;
+        cursor->err = MONGO_CURSOR_BSON_ERROR;
+        cursor->conn->err = MONGO_BSON_NOT_FINISHED;
         return MONGO_ERROR;
     }
 
     if( bson->err & BSON_NOT_UTF8 ) {
-        cursor->err = MONGO_BSON_INVALID;
+        cursor->err = MONGO_CURSOR_BSON_ERROR;
+        cursor->conn->err = MONGO_BSON_INVALID;
         return MONGO_ERROR;
     }
 
@@ -592,6 +591,7 @@ int mongo_update( mongo *conn, const char *ns, const bson *cond,
 
 int mongo_remove( mongo *conn, const char *ns, const bson *cond ) {
     char *data;
+    mongo_message *mm;
 
     /* Make sure that the BSON is valid UTF-8.
      * TODO: decide whether to check cond as well.
@@ -600,12 +600,12 @@ int mongo_remove( mongo *conn, const char *ns, const bson *cond ) {
         return MONGO_ERROR;
     }
 
-    mongo_message *mm = mongo_message_create( 16  /* header */
-                        + 4  /* ZERO */
-                        + strlen( ns ) + 1
-                        + 4  /* ZERO */
-                        + bson_size( cond )
-                        , 0 , 0 , MONGO_OP_DELETE );
+    mm = mongo_message_create( 16  /* header */
+                              + 4  /* ZERO */
+                              + strlen( ns ) + 1
+                              + 4  /* ZERO */
+                              + bson_size( cond )
+                              , 0 , 0 , MONGO_OP_DELETE );
 
     data = &mm->data;
     data = mongo_data_append32( data, &ZERO );
@@ -622,6 +622,15 @@ static int mongo_cursor_op_query( mongo_cursor *cursor ) {
     bson empty;
     char *data;
     mongo_message *mm;
+    bson temp;
+    bson_iterator it;
+
+    /* Clear any errors. */
+    bson_free( cursor->conn->lasterrstr );
+    cursor->conn->lasterrstr = NULL;
+    cursor->conn->lasterrcode = 0;
+    cursor->conn->err = 0;
+    cursor->err = 0;
 
     /* Set up default values for query and fields, if necessary. */
     if( ! cursor->query )
@@ -661,6 +670,19 @@ static int mongo_cursor_op_query( mongo_cursor *cursor ) {
     res = mongo_read_response( cursor->conn, ( mongo_reply ** )&( cursor->reply ) );
     if( res != MONGO_OK ) {
         return MONGO_ERROR;
+    }
+
+    if( cursor->reply->fields.num == 1 ) {
+        bson_init_data( &temp, &cursor->reply->objs );
+        if( bson_find( &it, &temp, "$err" ) ) {
+            cursor->conn->lasterrstr =
+              (char *)bson_malloc( bson_iterator_string_len( &it ) );
+            strcpy( cursor->conn->lasterrstr, bson_iterator_string( &it ) );
+            bson_find( &it, &temp, "code" );
+            cursor->conn->lasterrcode = bson_iterator_int( &it );
+            cursor->err = MONGO_CURSOR_QUERY_FAIL;
+            return MONGO_ERROR;
+        }
     }
 
     cursor->seen += cursor->reply->fields.num;
@@ -744,10 +766,17 @@ mongo_cursor *mongo_find( mongo *conn, const char *ns, bson *query,
 int mongo_find_one( mongo *conn, const char *ns, bson *query,
                     bson *fields, bson *out ) {
 
-    mongo_cursor *cursor = mongo_find( conn, ns, query, fields, 1, 0, 0 );
+    mongo_cursor cursor[1];
+    mongo_cursor_init( cursor, conn, ns );
+    mongo_cursor_set_query( cursor, query );
+    mongo_cursor_set_fields( cursor, fields );
+    mongo_cursor_set_limit( cursor, 1 );
 
-    if ( cursor && mongo_cursor_next( cursor ) == MONGO_OK ) {
-        bson_copy_basic( out, &cursor->current );
+    if ( mongo_cursor_next( cursor ) == MONGO_OK ) {
+        bson_init_size( out, bson_size( (bson *)&cursor->current ) );
+        memcpy( out->data, cursor->current.data,
+            bson_size( (bson *)&cursor->current ) );
+        out->finished = 1;
         mongo_cursor_destroy( cursor );
         return MONGO_OK;
     } else {
@@ -757,19 +786,11 @@ int mongo_find_one( mongo *conn, const char *ns, bson *query,
 }
 
 void mongo_cursor_init( mongo_cursor *cursor, mongo *conn, const char *ns ) {
+    memset( cursor, 0, sizeof( mongo_cursor ) );
     cursor->conn = conn;
     cursor->ns = ( const char * )bson_malloc( strlen( ns ) + 1 );
     strncpy( ( char * )cursor->ns, ns, strlen( ns ) + 1 );
     cursor->current.data = NULL;
-    cursor->reply = NULL;
-    cursor->flags = 0;
-    cursor->seen = 0;
-    cursor->err = 0;
-    cursor->options = 0;
-    cursor->query = NULL;
-    cursor->fields = NULL;
-    cursor->skip = 0;
-    cursor->limit = 0;
 }
 
 void mongo_cursor_set_query( mongo_cursor *cursor, bson *query ) {
@@ -805,7 +826,8 @@ int mongo_cursor_next( mongo_cursor *cursor ) {
     char *message_end;
 
     if( ! ( cursor->flags & MONGO_CURSOR_QUERY_SENT ) )
-        mongo_cursor_op_query( cursor );
+        if( mongo_cursor_op_query( cursor ) != MONGO_OK )
+            return MONGO_ERROR;
 
     if( !cursor->reply )
         return MONGO_ERROR;
@@ -827,7 +849,7 @@ int mongo_cursor_next( mongo_cursor *cursor ) {
 
     /* first */
     if ( cursor->current.data == NULL ) {
-        bson_init_data( &cursor->current, &cursor->reply->objs );
+        bson_init_finished_data( &cursor->current, &cursor->reply->objs );
         return MONGO_OK;
     }
 
@@ -844,9 +866,9 @@ int mongo_cursor_next( mongo_cursor *cursor ) {
             return MONGO_ERROR;
         }
 
-        bson_init_data( &cursor->current, &cursor->reply->objs );
+        bson_init_finished_data( &cursor->current, &cursor->reply->objs );
     } else {
-        bson_init_data( &cursor->current, next_object );
+        bson_init_finished_data( &cursor->current, next_object );
     }
 
     return MONGO_OK;
@@ -962,17 +984,34 @@ int64_t mongo_count( mongo *conn, const char *db, const char *ns, bson *query ) 
 int mongo_run_command( mongo *conn, const char *db, bson *command,
                        bson *out ) {
 
+    bson response = {NULL, 0};
     bson fields;
     int sl = strlen( db );
     char *ns = bson_malloc( sl + 5 + 1 ); /* ".$cmd" + nul */
-    int res;
+    int res, success = 0;
 
     strcpy( ns, db );
     strcpy( ns+sl, ".$cmd" );
 
-    res = mongo_find_one( conn, ns, command, bson_empty( &fields ), out );
+    res = mongo_find_one( conn, ns, command, bson_empty( &fields ), &response );
     bson_free( ns );
-    return res;
+
+    if( res != MONGO_OK )
+        return MONGO_ERROR;
+    else {
+        bson_iterator it;
+        if( bson_find( &it, &response, "ok" ) )
+            success = bson_iterator_bool( &it );
+
+        if( !success ) {
+            conn->err = MONGO_COMMAND_FAILED;
+            return MONGO_ERROR;
+        } else {
+            if( out )
+              *out = response;
+            return MONGO_OK;
+        }
+    }
 }
 
 int mongo_simple_int_command( mongo *conn, const char *db,
@@ -980,17 +1019,13 @@ int mongo_simple_int_command( mongo *conn, const char *db,
 
     bson out = {NULL, 0};
     bson cmd;
-    bson_bool_t success = 0;
+    int result;
 
     bson_init( &cmd );
     bson_append_int( &cmd, cmdstr, arg );
     bson_finish( &cmd );
 
-    if( mongo_run_command( conn, db, &cmd, &out ) == MONGO_OK ) {
-        bson_iterator it;
-        if( bson_find( &it, &out, "ok" ) )
-            success = bson_iterator_bool( &it );
-    }
+    result = mongo_run_command( conn, db, &cmd, &out );
 
     bson_destroy( &cmd );
 
@@ -999,30 +1034,21 @@ int mongo_simple_int_command( mongo *conn, const char *db,
     else
         bson_destroy( &out );
 
-    if( success )
-        return MONGO_OK;
-    else {
-        conn->err = MONGO_COMMAND_FAILED;
-        return MONGO_ERROR;
-    }
+    return result;
 }
 
 int mongo_simple_str_command( mongo *conn, const char *db,
                               const char *cmdstr, const char *arg, bson *realout ) {
 
     bson out = {NULL, 0};
-    int success = 0;
+    int result;
 
     bson cmd;
     bson_init( &cmd );
     bson_append_string( &cmd, cmdstr, arg );
     bson_finish( &cmd );
 
-    if( mongo_run_command( conn, db, &cmd, &out ) == MONGO_OK ) {
-        bson_iterator it;
-        if( bson_find( &it, &out, "ok" ) )
-            success = bson_iterator_bool( &it );
-    }
+    result = mongo_run_command( conn, db, &cmd, &out );
 
     bson_destroy( &cmd );
 
@@ -1031,10 +1057,7 @@ int mongo_simple_str_command( mongo *conn, const char *db,
     else
         bson_destroy( &out );
 
-    if( success )
-        return MONGO_OK;
-    else
-        return MONGO_ERROR;
+    return result;
 }
 
 int mongo_cmd_drop_db( mongo *conn, const char *db ) {
@@ -1170,7 +1193,7 @@ bson_bool_t mongo_cmd_authenticate( mongo *conn, const char *db, const char *use
     bson cmd;
     bson out;
     const char *nonce;
-    bson_bool_t success = 0;
+    int result;
 
     mongo_md5_state_t st;
     mongo_md5_byte_t digest[16];
@@ -1201,18 +1224,11 @@ bson_bool_t mongo_cmd_authenticate( mongo *conn, const char *db, const char *use
     bson_finish( &cmd );
 
     bson_destroy( &from_db );
-    /*bson_init( &from_db ); */
-    if( mongo_run_command( conn, db, &cmd, &out ) == MONGO_OK ) {
-        bson_iterator it;
-        if( bson_find( &it, &out, "ok" ) )
-            success = bson_iterator_bool( &it );
-    }
+
+    result = mongo_run_command( conn, db, &cmd, &out );
 
     bson_destroy( &from_db );
     bson_destroy( &cmd );
 
-    if( success )
-        return MONGO_OK;
-    else
-        return MONGO_ERROR;
+    return result;
 }